# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "API Documentation for the E-Commerce platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/admin/subcategories": {
            "post": {
                "tags": [
                    "Admin - Subcategories"
                ],
                "summary": "Create a new subcategory (Admin)",
                "description": "Add a new subcategory.",
                "operationId": "8db2d4ff908a186cda7c27f0fb879e24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/subcategories/{id}": {
            "put": {
                "tags": [
                    "Admin - Subcategories"
                ],
                "summary": "Update a subcategory (Admin)",
                "description": "Update an existing subcategory.",
                "operationId": "eb7cf94446abbcdba78d4f890d592085",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subcategory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Subcategories"
                ],
                "summary": "Delete a subcategory (Admin)",
                "description": "Delete a subcategory from the database.",
                "operationId": "dc4f0386fa6ab6e78c610992b2db6d78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subcategory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subcategory deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Get user cart",
                "description": "Retrieve the current user's cart and its items.",
                "operationId": "2e9cdf3609eb06e5e6215892aa750e4c",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/items": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "Add item to cart",
                "description": "Add a new product to the user's cart.",
                "operationId": "bc4f11f34c70d4777642a8e57e5169c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "add_ons": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item added successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/items/{id}": {
            "put": {
                "tags": [
                    "Carts"
                ],
                "summary": "Update cart item quantity",
                "description": "Update the quantity of an existing item in the cart.",
                "operationId": "95a77f39e5ad6e76f57340381bb7ba31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cart Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "Remove item from cart",
                "description": "Remove a specific item from the user's cart.",
                "operationId": "7eb8301f29ac5e1266b6e69744cc26f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cart Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removed successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}